backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic-settings==2.5.2
slowapi==0.1.9
sqlalchemy==2.0.35
alembic==1.18.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
